Make sure admins are auth{enticated,orized} before accessing promotions endpoints

+++ backend/src/main/java/com/cinema/backend/controllers/PromotionsController.java
@@ -4,10 +4,10 @@ import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 import com.cinema.backend.entities.Promotion;
 import com.cinema.backend.repositories.PromotionsRepository;
+import com.cinema.backend.services.AccountsService;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -24,36 +24,34 @@ import org.springframework.web.server.ResponseStatusException;
 public class PromotionsController {
 
   private final PromotionsRepository promotionsRepository;
+  private final AccountsService accountsService;
+  private final HttpServletRequest request;
 
   @Autowired
-  public PromotionsController(PromotionsRepository promotionsRepository) {
+  public PromotionsController(
+      PromotionsRepository promotionsRepository,
+      AccountsService accountsService,
+      HttpServletRequest request) {
     this.promotionsRepository = promotionsRepository;
+    this.accountsService = accountsService;
+    this.request = request;
   }
 
   @PostMapping("/create")
-  public Map<String, String> createPromotion(@Valid @RequestBody Promotion promotion) {
+  public void createPromotion(@Valid @RequestBody Promotion promotion) {
+    accountsService.ensureAdmin(request);
     promotionsRepository.save(promotion);
-    return new HashMap<>() {
-      {
-        this.put("id", promotion.code);
-      }
-    };
-  }
-
-  @GetMapping("/get/{id}")
-  public Promotion getPromotion(@PathVariable long id) {
-    return promotionsRepository
-        .findById(id)
-        .orElseThrow(() -> new ResponseStatusException(NOT_FOUND));
   }
 
   @GetMapping("/getAll")
   public List<Promotion> getPromotions() {
+    accountsService.ensureAdmin(request);
     return promotionsRepository.findAll();
   }
 
   @PutMapping("/update/{id}")
   public void updatePaymentCard(@PathVariable long id, @Valid @RequestBody Promotion promotion) {
+    accountsService.ensureAdmin(request);
     var dbPromotion =
         promotionsRepository.findById(id).orElseThrow(() -> new ResponseStatusException(NOT_FOUND));
     dbPromotion.copy(promotion);
@@ -62,6 +60,7 @@ public class PromotionsController {
 
   @DeleteMapping("/delete/{id}")
   public void deletePaymentCard(@PathVariable long id) {
+    accountsService.ensureAdmin(request);
     promotionsRepository.deleteById(id);
   }
 }
